### 2018-01-18 #539 @arfon
👋 @julia-wrobel - it looks like there are some issues with the YAML header in your `paper.md` file. Could you modify it so it looks like this: https://gist.github.com/arfon/457fb87bdea6285b53bd23480bfe357a ?
