### 2018-01-18 6750dc8
formatting changes to YAML header for paper.md
