### 2018-01-18 #539 @julia-wrobel
@arfon - I modified the YAML header in the paper.md file. Thanks! Let me know if I need additional changes.
