### 2018-01-18 452514b
edit equation in paper
