### 2018-01-18 #539 @arfon
@julia-wrobel - it looks like Pandoc doesn't like the math syntax that you have. Can you try modifying it to look like this:

```
$$
\begin{aligned}
E\left[Y_i\left(h_i^{-1}(t_i^*)\right) | c_i, h_i^{-1} \right] &=& \mu_i(t)
g\left[\mu_i(t)\right]&=& \alpha(t) + \sum_{k = 1}^K c_{ik}\psi_k(t).
\end{aligned}
$$
```
