### 2018-01-18 9bee0ea
edits to paper.md equaiton
