### 2018-01-18 #539 @julia-wrobel
Thanks @arfon! I just changed it - instead of aligning the equations I just separated them into two. Hopefully that should fix the syntax error!
