### 2018-01-28 #539 @arfon
@corybrunson - would you be willing to review this submission for JOSS?
