### 2018-01-29 #539 @corybrunson
@arfon – yes, i'd be glad to! I will need to refresh on functional analysis and familiarize myself with the JOSS process, and i can say later today if i don't feel qualified. Thank you.
