### 2018-01-29 #539 @arfon
> Edit: The technique and implementation are accessible enough to me that i believe i can provide a useful review.

Excellent. Thanks! I'll go ahead and start setting up the review now.
