### 2018-01-29 #557 @whedon
[REVIEW]: registr: Registration for Exponential Family Functional Data
**Submitting author:** @julia-wrobel (<a href="http://orcid.org/0000-0001-6783-1421">Julia Wrobel</a>)
**Repository:** <a href="https://github.com/julia-wrobel/registr" target ="_blank">https://github.com/julia-wrobel/registr</a>
**Version:** 0.1
