### 2018-01-29 #539 @arfon
@corybrunson - please head over to #557 for the main review issue.
