### 2018-02-01 julia-wrobel/registr#2 @corybrunson
vignette error
On knitting the vignette i hit a small error, which i've summarized below (with my own `sessionInfo()`):

``` r
library(registr)
library(dplyr)
#>
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#>
#>     filter, lag
#> The following objects are masked from 'package:base':
#>
#>     intersect, setdiff, setequal, union
library(ggplot2)
sessionInfo()
#> R version 3.3.2 (2016-10-31)
#> Platform: x86_64-apple-darwin13.4.0 (64-bit)
#> Running under: OS X Mavericks 10.9.5
#>
#> locale:
#> [1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
#>
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base
#>
#> other attached packages:
#> [1] ggplot2_2.2.1.9000 dplyr_0.7.4        registr_0.1
#>
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_0.12.15      knitr_1.18        bindr_0.1
#>  [4] magrittr_1.5      splines_3.3.2     munsell_0.4.3
#>  [7] colorspace_1.3-2  R6_2.2.2          rlang_0.1.6.9003
#> [10] plyr_1.8.4        stringr_1.2.0     tools_3.3.2
#> [13] grid_3.3.2        gtable_0.2.0      withr_2.1.1.9000
#> [16] htmltools_0.3.6   lazyeval_0.2.1    yaml_2.1.16
#> [19] assertthat_0.2.0  rprojroot_1.3-2   digest_0.6.14
#> [22] tibble_1.4.2      bindrcpp_0.2      purrr_0.2.4
#> [25] tidyr_0.7.2       glue_1.2.0        evaluate_0.10.1
#> [28] rmarkdown_1.8     stringi_1.1.6     pillar_1.1.0
#> [31] scales_0.5.0.9000 backports_1.1.2   boot_1.3-20
#> [34] pkgconfig_2.0.1
fpca_data = simulate_functional_data(I = 100, D = 200)
Y = fpca_data$Y
pc_df = data.frame(pop_mean = fpca_data$alpha,
                   psi1 = fpca_data$psi1,
                   psi2 = fpca_data$psi2,
                   index = seq(0, 1, length.out = 200),
                   id = 1)
bfpca_object = bfpca(fpca_data$Y, npc = 2, Kt = 8, print.iter = TRUE)
#> current iteration: 1
#> current error: 100
#> current iteration: 2
#> current error: 7.46139390691072
#> current iteration: 3
#> current error: 5.36360041515815
#> current iteration: 4
#> current error: 0.736588648381846
#> current iteration: 5
#> current error: 0.141123511845482
#> current iteration: 6
#> current error: 0.0528979518549368
#> current iteration: 7
#> current error: 0.0261549215886246
#> current iteration: 8
#> current error: 0.0152654597821408
#> current iteration: 9
#> current error: 0.0099486416630379
#> current iteration: 10
#> current error: 0.00699015439921455
#> current iteration: 11
#> current error: 0.00517540083860305
#> current iteration: 12
#> current error: 0.00397781621072642
#> current iteration: 13
#> current error: 0.00314286307931933
#> current iteration: 14
#> current error: 0.00253598531387938
#> current iteration: 15
#> current error: 0.00208054812699585
#> current iteration: 16
#> current error: 0.00173010167570517
#> current iteration: 17
#> current error: 0.00145499388634334
#> current iteration: 18
#> current error: 0.00123546459162101
#> current iteration: 19
#> current error: 0.00105786936190601
pc_df = pc_df %>%
  mutate(psi1_est = bfpca_object$efunctions[,1],
         psi2_est = bfpca_object$efunctions[,2],
         alpha_est = bfpca_object$alpha)
ggplot(pc_df, aes(index, pop_mean)) + theme_bw() + geom_line(color = "blue") +
  geom_line(aes(y = alpha_est), linetype = 2, color = "red")
#> Error: Column `y` must be a 1d atomic vector or a list
```

![](https://i.imgur.com/2fwH6K9.png)

The object `bfpca_object$alpha` is a 200-by-1 matrix, consistent with the `bfpca` tests, so using `alpha_est = as.vector(bfpca_object$alpha)` resolved it for me; i'm not sure if other functionality might be affected, but the vignette runs fine otherwise!
