### 2018-02-02 #557 @corybrunson
A small error in the vignette ([issue raised at the submitted repository](https://github.com/julia-wrobel/registr/issues/2)) needs to be resolved in order for the package to install (with `build_vignettes = TRUE`).
