### 2018-02-02 #557 @corybrunson
@arfon may i ask questions of the submitting author here, or should all feedback to them be at the submitted repository? And at what point—and how—should i grade the software into a category (e.g. "Minor Revisions")?
