### 2018-02-02 2cf7775
fixed build error for vignette when devel version of ggplot2 is loaded
