### 2018-02-02 julia-wrobel/registr#2 @julia-wrobel
Hey Cory - I was able to reproduce your error - it seems that the vignette build was incompatible with the devel version of ggplot2 due to a way I defined one of the dataframes. I fixed the error by converting alpha_est to a matrix as you noticed worked - thanks! You should  be able to build the vignette now.
