### 2018-02-02 julia-wrobel/registr#2 @corybrunson
Successfully re-installed. : )
