### 2018-02-03 6a29045
update documentation for Kt and Kh parameters
