+++ R/bfpca.R
@@ -5,7 +5,7 @@
 #' binary functional data.
 #'
 #' @param Y Dataframe. Should have variables id, value, index. 
-#' @param npc Defaults to 1. Number of principal components to calculate.
+#' @param npc Default is 1. Number of principal components to calculate.
 #' @param Kt Number of B-spline basis functions used to estimate mean functions. Default is 8.
 #' @param maxiter Maximum number of iterations to perform for EM algorithm. Default is 50.
 #' @param t_min Minimum value to be evaluated on the time domain. 
