+++ README.md
@@ -57,11 +57,11 @@ registration_data = simulate_unregistered_curves(I = 100, D = 200, seed = 2018)
 
 The plot below shows the unregistered curves and registered curves.
 
-<img src="README_files/figure-markdown_github-ascii_identifiers/plot_sim_data-1.png" style="display: block; margin: auto;" />
+<img src="README_files/figure-markdown_github/plot_sim_data-1.png" style="display: block; margin: auto;" />
 
 Continuously observed curves are shown above in order to illustrate the misalignment problem and our simulated data; the simulated dataset also includes binary values which have been generated by using these continuous curves as probabilities. The unregistered and registered binary curves for two subjects are shown below.
 
-<img src="README_files/figure-markdown_github-ascii_identifiers/plot_2subjs-1.png" style="display: block; margin: auto;" />
+<img src="README_files/figure-markdown_github/plot_2subjs-1.png" style="display: block; margin: auto;" />
 
 Our software registers curves by estimating *t*. For this we use the function `registration_fpca()`.
 
@@ -73,9 +73,9 @@ binary_registration = register_fpca(Y = registration_data, family = "binomial",
 ## current iteration: 3
 ```
 
-The plot below shows unregistered, true registered, and estimated registered binary curves two subjects after fitting our method.
+The plot below shows unregistered, true registered, and estimated registered binary curves for two subjects after fitting our method.
 
-<img src="README_files/figure-markdown_github-ascii_identifiers/plot_fit-1.png" style="display: block; margin: auto;" />
+<img src="README_files/figure-markdown_github/plot_fit-1.png" style="display: block; margin: auto;" />
 
 ### Contributions
 
