### 2018-02-03 5e3a327
removed as_refund_obj functions from package and documentation
