+++ NAMESPACE
@@ -1,7 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(as_refundObj,matrix)
-export(as_refundObj)
 export(bfpca)
 export(bs_deriv)
 export(constraints)
@@ -26,14 +24,12 @@ importFrom(magrittr,"%>%")
 importFrom(splines,bs)
 importFrom(splines,splineDesign)
 importFrom(stats,coef)
-importFrom(stats,complete.cases)
 importFrom(stats,constrOptim)
 importFrom(stats,glm)
 importFrom(stats,quantile)
 importFrom(stats,rbinom)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
-importFrom(stats,setNames)
 importFrom(tidyr,gather)
 importFrom(tidyr,spread)
 useDynLib(registr)
