### 2018-02-06 julia-wrobel/registr#3 @corybrunson
boot dependency
**registr** depends on **boot**, apparently only for `boot::inv.logit()`, which itself is a wrapper for `plogis()`. In the interest of reducing unnecessary dependencies, would it affect the functionality to drop **boot**?

(I wouldn't hold up acceptance for this issue, but it seems worth raising.)

_This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/557)._
