### 2018-02-06 julia-wrobel/registr#5 @corybrunson
link FPCA and registration function documentation
This is a minor issue, but it slowed me down a bit: The documentation for `register_fpca()` mentions the FPCA step but doesn't say how that step is performed. I think it would be helpful to include mentions of / links to the binary and Gaussian FPCA functions, e.g. `\code{\link{bfpca}}`, and perhaps similar links elsewhere, so that the user can trace their issues through the documentation more easily. (Won't hold up my recommendation for acceptance but wanted to point it out.)

_This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/557)._
