### 2018-02-10 b3e668a
removed boot dependency by replacing inv.logit() with plogis()
