### 2018-02-10 julia-wrobel/registr#3 @julia-wrobel
Good call!  I dropped the boot dependency and replaced inv.logit() with plogis()
