### 2018-02-10 bcad10e
update documentation so that functions link to other functions
