+++ man/register_fpca.Rd
@@ -37,9 +37,15 @@ family \code{gaussian} or \code{binomial}.
 }
 \description{
 Function combines constrained optimization and FPCA to estimate warping functions for 
-exponential family curves. Requires input data \code{Y} to be a dataframe in long format with variables 
+exponential family curves. The FPCA step is performed through the function 
+\code{\link{bfpca}} if \code{family = "binomial"} or the function 
+\code{\link{fpca_gauss}} if \code{family = "gaussian"}. Warping functions are calculated 
+by the function \code{\link{registr}}.
+}
+\details{
+Requires input data \code{Y} to be a dataframe in long format with variables 
 \code{id}, \code{index}, and \code{value} to indicate subject IDs, times, and observations, 
-respectively.
+respectively. The code calls two
 }
 \examples{
 
