+++ man/fpca_gauss.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/fpca_gauss.R
 \name{fpca_gauss}
 \alias{fpca_gauss}
-\title{Functional principal components analysis via variational bayes}
+\title{Functional principal components analysis via variational EM}
 \usage{
 fpca_gauss(Y, npc = 1, Kt = 8, maxiter = 20, t_min = NULL,
   t_max = NULL, print.iter = FALSE, row_obj = NULL, seed = 1988, ...)
@@ -47,17 +47,14 @@ For use in \code{registr()} function.}
 \item{sigma2}{Estimated error variance}
 }
 \description{
-Function used in the FPCA step for registering functional data when \code{family = "gaussian"}. 
-This method uses mean field variational inference to estimate scores and principal components for 
-continuous functional data.
+Function used in the FPCA step for registering functional data,
+called by \code{\link{register_fpca}} when \code{family = "gaussian"}. 
+Parameters estimated based on probabilistic PCA framework orginally 
+introduced by Tipping and Bishop in 1999.
 }
 \references{
-Jaakkola, T. S. and Jordan, M. I. (1997).
-A variational approach to Bayesian logistic regression models and their extensions. 
-\emph{A variational approach to Bayesian logistic regression models and their extensions.}.
-
-Tipping, M. E. (1999). Probabilistic Visualisation of High-dimensional binary data.
-\emph{Advances in neural information processing systems}, 592--598.
+Tipping, M. E. and Bishop, C (1999). Probabilistic Principal Component Analysis.
+\emph{Journal of the Royal Statistical Society Series B,}, 592--598.
 }
 \author{
 Julia Wrobel \email{jw3134@cumc.columbia.edu},
