### 2018-02-12 7aede8d
updated example in register_fpca documentation
