+++ R/register_fpca.R
@@ -37,7 +37,7 @@
 #'
 #' \dontrun{
 #'  Y = simulate_unregistered_curves(I = 50, D = 200)
-#'  registr_object = register_fpca(Y, family = "binomial", max_iterations = 25)
+#'  registr_object = register_fpca(Y, family = "binomial", max_iterations = 5)
 #' }
 #'
 register_fpca <- function(Y, Kt = 8, Kh = 4, family = "binomial", max_iterations = 10, npc = 1, ...){
