+++ man/register_fpca.Rd
@@ -51,7 +51,7 @@ respectively. The code calls two
 
 \dontrun{
  Y = simulate_unregistered_curves(I = 50, D = 200)
- registr_object = register_fpca(Y, family = "binomial", max_iterations = 25)
+ registr_object = register_fpca(Y, family = "binomial", max_iterations = 5)
 }
 
 }
