+++ man/register_fpca.Rd
@@ -52,6 +52,10 @@ respectively. The code calls two
 \dontrun{
  Y = simulate_unregistered_curves(I = 50, D = 200)
  registr_object = register_fpca(Y, family = "binomial", max_iterations = 5)
+ 
+ # example using accelerometer data from nhanes 2003-2004 study
+ data(nhanes)
+ register_nhanes = register_fpca(nhanes, npc = 2, family = "binomial", max_iterations = 5)
 }
 
 }
