### 2018-02-16 julia-wrobel/registr#4 @julia-wrobel
I added a data folder and in it included a subset of accelerometer data from the publicly available national health and nutrition examination survey (NHANES) data. This is the same style of data that we used in our methods paper and I processed it so that it works with our software. I also included an example using this dataset in the documentation for the register_fpca() function. Does that work?
