### 2018-02-16 #557 @corybrunson
@arfon the issues i raised have been addressed, and i recommend acceptance of the current version. Thank you!
