### 2018-02-17 #557 @arfon
> @arfon the issues i raised have been addressed, and i recommend acceptance of the current version. Thank you!

👍 great!
